Added the --verbose output in case of ICMP errors

+++ include/config.h
@@ -2,12 +2,13 @@
 
 #include <unistd.h>
 #include <stdint.h>
+#include <arpa/inet.h>
 
 typedef struct s_config {
 	uint8_t		verbose;
 	uint8_t		quiet;
 	uint16_t	identifier;
 	size_t		payload_size;
-	uint8_t		count;
-	uint8_t		interval;
+	ssize_t		count;
+	ssize_t		interval;
 } t_config;
